Share one Westwood delta decoder between AUD and VQA audio
